% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fips.R, R/utils.R
\name{create_fips}
\alias{create_fips}
\title{This function takes in a state vector, county vector, and tract vector
that are string or numeric, and returns a unified FIPS code of the finest
level of geography provided}
\usage{
create_fips(state = NULL, county = NULL, tract = NULL)

create_fips(state = NULL, county = NULL, tract = NULL)
}
\description{
This function takes in a state vector, county vector, and tract vector
that are string or numeric, and returns a unified FIPS code of the finest
level of geography provided

Helper function for smoov(). This function accepts input for states,
counties, and tracts, and outputs an vector equal to the length of each
input vector, or the length of the smallest geography vector if only
one of each of the larger geographies is provided.
}
