% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_smoov_plot.R
\name{shp_data_merge}
\alias{shp_data_merge}
\title{Helper function for smoov()
This function takes in map data and the dataframe of values, merging them
together when appropriate. This function also performs the subsetting
operations if subsetting parameters (states, counties, tracts) are supplied.
shp_data_merge() returns a logical vector of length three that
1. Should all of USA be plotted? TRUE/FALSE
2. Should Hawaii be plotted? TRUE/FALSE
3. Should Alaska be plotted? TRUE/FALSE}
\usage{
shp_data_merge(shp, data = NULL, id = "fips", value = NULL, subfips = NULL)
}
\description{
Helper function for smoov()
This function takes in map data and the dataframe of values, merging them
together when appropriate. This function also performs the subsetting
operations if subsetting parameters (states, counties, tracts) are supplied.
shp_data_merge() returns a logical vector of length three that
1. Should all of USA be plotted? TRUE/FALSE
2. Should Hawaii be plotted? TRUE/FALSE
3. Should Alaska be plotted? TRUE/FALSE
}
