% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoov_border.R
\name{smoov_border}
\alias{smoov_border}
\title{Add border layers to plot}
\usage{
smoov_border(
  geo = NULL,
  line_color = "black",
  line_size = 0.1,
  line_alpha = 1,
  fill_color = NA,
  fill_alpha = 0.5,
  states = NULL,
  counties = NULL,
  tracts = NULL,
  year = 2010,
  detailed = "coarse",
  class = "sf"
)
}
\arguments{
\item{geo}{String. May include: "tracts", "counties", or
"states"}

\item{line_color}{String. Color of borders. Default is "black"}

\item{line_size}{Numeric. Default is 0.1}

\item{line_alpha}{Numeric in \link{0,1}. Transparency of borders.0 for completely transparent
and 1 for completely opaque}

\item{fill_color}{String. Color of fill within borders. Default is NA (i.e. transparent)}

\item{fill_alpha}{String. Transparency of fill within borders. Default is 0.5}

\item{...}{Optional subsetting parameters and shapefile options. See \code{?smoov}}
}
\description{
Add border layers to plot
}
