% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoov_plot.R
\name{smoov_plot}
\alias{smoov_plot}
\title{Helper function for \code{smoov()}}
\usage{
smoov_plot(
  geo,
  data,
  value,
  year,
  detailed,
  class,
  states = NULL,
  tracts = NULL,
  counties = NULL
)
}
\description{
Helper function for \code{smoov()}
}
