% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoov_setup.R
\name{smoov_setup}
\alias{smoov_setup}
\title{Simple mapping tool}
\usage{
smoov_setup(load_mapfiles = FALSE, load_us_geos = FALSE, load_roads = FALSE)
}
\arguments{
\item{load_mapfiles}{logical; FALSE by default. If TRUE, loads and prepares
mapfiles for use in the \code{smoov()} function. If only certain geographies
wish to be loaded, use instead \code{load_us_geos}, \code{load_roads}, or
\code{load_sch_districts}.}

\item{load_us_geos}{logical; FALSE by default. If TRUE then loads US tracts,
counties, and states.}

\item{load_roads}{logical; FALSE by default. If TRUE then loads US roads
for counties of the top 25 US cities by population}
}
\description{
Simple mapping tool
}
