% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoov.R
\name{smoov}
\alias{smoov}
\title{Simple mapping tool}
\usage{
smoov(
  geo,
  data = NULL,
  value = NULL,
  id = "fips",
  year = 2010,
  detailed = FALSE,
  class = "sf",
  states = NULL,
  counties = NULL,
  tracts = NULL,
  ...
)
}
\arguments{
\item{geo}{Character string; Provide level of map geography e.g. county.
See details.}

\item{data}{Dataframe; ID columns matching the level of geography. See details.}

\item{value}{character; Name of the column in \code{data} with which a
choropleth is created. If no column is provided, a map of boundaries
is produced. If more than one value is inputted, a faceted map is
created.}

\item{id}{character; "fips" by default.
Name of the column in \code{data} that provides the relevant
geography identifier (e.g. FIPS code).}

\item{year}{integer; Currently either 2000 or 2010. Year of census geography. The
most recent year by default.}

\item{detailed}{logical; FALSE by default. If TRUE, then  more detailed geographies are
used when available}

\item{class}{\code{"sp"} or \code{"sf"}; \code{"sf"} by default.}

\item{...}{Optional aesthetic parameters, subsetting parameters. See details below.}
}
\value{
\preformatted{Output will be a `ggplot` base layer onto which any `ggplot` layer may
be added. 
}
}
\description{
Simple mapping tool
}
\section{Geographic Subsetting Parameters}{

\itemize{
\item \code{states}: Vector of state \href{https://en.wikipedia.org/wiki/Federal_Information_Processing_Standard_state_code}{FIPS} codes.
Character and numeric vectors allowed.
\item \code{counties}: Vector of county \href{https://www.nrcs.usda.gov/wps/portal/nrcs/detail/national/home/?cid=nrcs143_013697}{FIPS} codes.
Character and numeric vectors allowed. If length of \code{states} vector is
longer than one, then \code{counties} vector must be of the same length.
\item \code{tracts}: Vector of tract \href{https://www.policymap.com/2012/08/tips-on-fips-a-quick-guide-to-geographic-place-codes-part-iii/}{FIPS} codes.
Character and numeric vectors allowed. If length of \code{states} vector or
\code{counties} vector islonger than one, then \code{tracts} vector must be of
the same length as \code{counties} vector.
}
}

\section{Aesthetic Parameters}{

\itemize{
\item \code{gradient}: The color gradient to be plotted, currently only "redblue"
\item \code{gradient_dir}: The direction the color gradient should go. Default is
\code{1} for low values to get the first value of the gradient and
\code{-1} is for low values to get the second value of the
color gradient.
\item \code{gradient_breaks}: A vector of five probabilities in \eqn{[0,1]^5}
that define the quantiles of \code{values} that the gradient emphasizes.
Defaults to \code{c(0,0.35,0.5,0.65,1)}
\item \code{linesize}: The size of the borders of geographic units. Defaults to
0.05 for plotting tracts, 0.1 for other levels of geography
\item \code{alpha}: Transparency parameter for fill colors. 0 is completely
transparent and 1 is completely opaque. Defaults to 0.9
\item \code{bordercolor}: Color of geography borders
}
}

