% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoov_plot.R
\name{smoov_plot}
\alias{smoov_plot}
\title{Helper function for \code{smoov()}}
\usage{
smoov_plot(
  geo,
  data,
  value = NULL,
  id,
  year,
  detailed,
  class,
  subsetfips = NULL,
  gradient = "redblue",
  gradient_dir = 1,
  gradient_breaks = c(0, 0.3, 0.5, 0.7, 1),
  linesize = 0.1,
  alpha = 0.9,
  bordercolor = NA,
  legend_sigfigs = 2
)
}
\description{
Helper function for \code{smoov()}
}
